+++ .scripts/ci/create_testing_project.sh
@@ -5,7 +5,7 @@ echo "Creating Testing Project"
 PACKAGE_FOLDER=${UNITY_DIR}/Assets/ElympicsSDK/
 
 ${UNITY_EXECUTABLE:-xvfb-run --auto-servernum --server-args='-screen 0 640x480x24' unity-editor} \
-  -createProject $UNITY_DIR \
+  -createProject "$UNITY_DIR" \
   -logFile /dev/stdout \
   -batchmode \
   -nographics \
@@ -22,9 +22,9 @@ fi
 
 echo "Moving Package files to appropriate directories"
 
-mkdir -p ${PACKAGE_FOLDER}
-cp -r Editor ${PACKAGE_FOLDER}
-cp -r Runtime ${PACKAGE_FOLDER}
-cp -r Tests ${PACKAGE_FOLDER}
+mkdir -p "$PACKAGE_FOLDER"
+cp -r Editor "$PACKAGE_FOLDER"
+cp -r Runtime "$PACKAGE_FOLDER"
+cp -r Tests "$PACKAGE_FOLDER"
 
 echo "Elympics moved ✅"