+++ Editor/AssemblyInfo.cs
@@ -4,6 +4,4 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyCompany("Elympics")]
 [assembly: AssemblyProduct("Elympics SDK")]
 [assembly: AssemblyVersion("0.7.3.0")]
-[assembly: InternalsVisibleTo("Elympics.Tests")]
-[assembly: InternalsVisibleTo("Elympics.Editor")]
 [assembly: InternalsVisibleTo("Elympics.Editor.Tests")]